Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.28.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
pydantic==2.7.1